修复BUG：DgvSensor的cb框提示信息不匹配，dgvname的状态显示FFFFFFFF出现异常（不能控制）。完成逻辑场景识别的窗体数据展现部分

+++ eNet编辑器/DgvView/DgvName.cs
@@ -1080,7 +1080,7 @@ namespace eNet编辑器.DgvView
                 
             }
             //再反转复原二进制数据
-            return Convert.ToInt32(DataChange.Reversal(bin), 2).ToString();
+            return Convert.ToInt64(DataChange.Reversal(bin), 2).ToString();
         }
 
         #endregion
